+++ Modular_Arithmetic/TestModulo.c
@@ -14,7 +14,7 @@
  * @param m the module value.
  * @return 1 if the 2 number are congruent, 0 otherwise.
  */
-int areCongruent(int a, int b, int m) {
+long long int areCongruent(long long int a, long long int b, long long int m) {
     return a % m == b % m;
 }
 
@@ -26,7 +26,7 @@ int areCongruent(int a, int b, int m) {
  * @param n the second number.
  * @return 1 if the 2 number are coprime, 0 otherwise.
  */
-int areCoPrime(int a, int n) {
+long long int areCoPrime(long long int a, long long int n) {
     return gcd(a, n) == 1;
 }
 
@@ -37,7 +37,7 @@ int areCoPrime(int a, int n) {
  * @param m the second number.
  * @return 1 if the first number is a divisor of the second number, 0 otherwise.
  */
-int isDivisor(int n, int m) {
+long long int isDivisor(long long int n, long long int m) {
     return m % n == 0;
 }
 
@@ -53,7 +53,7 @@ int isDivisor(int n, int m) {
  * @param n the second number.
  * @returns 1 if the number is Fermat's Pseuodprime to a, 0 otherwise.
  */
-int isFermatPseudoPrime(int a, int n) {
+long long int isFermatPseudoPrime(long long int a, long long int n) {
     assert(areCoPrime(a, n) == 1 && "a must be coprime with n");
 
     return power(a, n - 1, n) == 1;
@@ -66,33 +66,33 @@ int isFermatPseudoPrime(int a, int n) {
  * @param n the number whose primality is to be tested.
  * @return 1 if the number is prime, 0 otherwise.
  */
-int isPrime(int n) {
+long long int isPrime(long long int n) {
     if (n == 0 || n == 1) {
         return 0;
     }
 
     //The list of all prime numbers.
-    int *primes = malloc((n + 1) * sizeof(int));
+    long long int *primes = malloc((n + 1) * sizeof(int));
     //The index of the last prime number found.
-    int primeIndex = 0;
+    long long int primeIndex = 0;
     //The list of all number until n.
-    int *numbers = malloc((n + 1) * sizeof(int));
+    long long int *numbers = malloc((n + 1) * sizeof(int));
 
     //array initialization: assume that all numbers are primes
-    for (int i = 0; i < n + 1; ++i) {
+    for (long long int i = 0; i < n + 1; ++i) {
         numbers[i] = 1;
     }
     numbers[0] = 0;
     numbers[1] = 0;
 
-    for (int i = 2; (i * i) < ((n + 1) * (n + 1)); ++i) {
+    for (long long int i = 2; (i * i) < ((n + 1) * (n + 1)); ++i) {
         //check if i is prime
         if (numbers[i]) {
             //i is a prime number
             primes[primeIndex++] = i;
 
             //updates all multiples of i as non-prime
-            for (int j = (i * i); j < (n + 1); j += i) {
+            for (long long int j = (i * i); j < (n + 1); j += i) {
                 numbers[j] = 0;
                 if (j == n) {
                     free(numbers);
@@ -118,7 +118,7 @@ int isPrime(int n) {
  * @param n the module value.
  * @return 1 if the number allows for the square root, 0 otherwise.
  */
-int isSquareNumber(int a, int n) {
+long long int isSquareNumber(long long int a, long long int n) {
     a = mod(a, n);
 
     //Euler's criterion
@@ -126,7 +126,7 @@ int isSquareNumber(int a, int n) {
         return power(a, (int) ((n - 1) / 2), n) == 1;
     }
 
-    for (int i = 0; i < (int) ((n - 1) / 2); ++i) {
+    for (long long int i = 0; i < (int) ((n - 1) / 2); ++i) {
         if (product(i, i, n) == a) {
             return 1;
         }
@@ -142,15 +142,15 @@ int isSquareNumber(int a, int n) {
  * @param n the module value.
  * @return 1 if the number is a primitive root modulo n, 0 otherwise.
  */
-int isPrimitiveRoot(int a, int n) {
+long long int isPrimitiveRoot(long long int a, long long int n) {
     //The value of Euler function of n.
-    int phi = EulerFunction(n);
+    long long int phi = EulerFunction(n);
     //The number of factor.
-    int factorPhiSize = 0;
+    long long int factorPhiSize = 0;
     //The list of factors of phi.
-    int *factorsPhi = factorisation(phi, &factorPhiSize);
+    long long int *factorsPhi = factorisation(phi, &factorPhiSize);
 
-    for (int i = 0; i < factorPhiSize; ++i) {
+    for (long long int i = 0; i < factorPhiSize; ++i) {
         if (power(a, (int) (phi / factorsPhi[i]), n) == 1) {
             free(factorsPhi);
             return 0;
